MKTAXI-3987 => Ensuring tests accounts continues

+++ Src/Server/MK.Booking.Api/Controllers/Api/Admin/AdminCreditCardController.cs
@@ -11,7 +11,6 @@ using Infrastructure.Messaging;
 
 namespace apcurium.MK.Web.Controllers.Api.Admin
 {
-    [RoutePrefix("api/v2/admin")]
     public class AdminCreditCardController : BaseApiController
     {
         private readonly CreditCardService _creditCardService;
@@ -28,7 +27,7 @@ namespace apcurium.MK.Web.Controllers.Api.Admin
             PrepareApiServices(_creditCardService);
         }
 
-        [HttpDelete, Auth(Role = RoleName.Support), Route("deleteAllCreditCards/{accountId}")]
+        [HttpDelete, Auth(Role = RoleName.Support), Route("api/v2/admin/deleteAllCreditCards/{accountId}")]
         public IHttpActionResult Delete(Guid accountId)
         {
             _creditCardService.Delete(new DeleteCreditCardsWithAccountRequest() {AccountID = accountId});
